chore(api/applications): rename published-fields-changed to specify case
